requirements.txt
google-cloud-logging==1.10.0
google-cloud-pubsub==0.39.1
websocket==0.2.1
websocket-client==0.55.0
redis==3.2.0
cerberus==1.2
apscheduler==3.5.3
python-dateutil==2.8.0
fluent-logger==0.9.3
prometheus-client==0.6.0
mysql-connector-python==8.0.16